+++ packages/components/drawer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sl-design-system/drawer
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.8
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes